changed datatype for some reason

+++ heavenLi_pyopengl/hliGLutils/vertAttribStruct.h
@@ -23,7 +23,7 @@ class vertAttribMetadata {
      
    //private:
    protected:
-      GLubyte  vectorSize  = 0,  // Number of vector components (1-4)
+      GLuint   vectorSize  = 0,  // Number of vector components (1-4)
                vectorOffset= 0;  // May get deprecated
       GLuint   attribIndex = 0;  // May get deprecated
 };
